# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateCoupleRequest {
  /**
     * @minLength 1
     * @maxLength 80
     */
  userName: string;
}

export interface CreateSessionRequest {
  /**
     * @minLength 8
     * @maxLength 40
     */
  accountKey: string;
  /**
     * @minLength 6
     * @maxLength 64
     */
  recoveryPin: string;
  /** @maxLength 80 */
  userName?: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  birthday?: string;
  /** @maxLength 40 */
  pronouns?: string;
}

export interface UserProfile {
  id: string;
  name: string;
  birthday?: string;
  pronouns?: string;
}

export interface CoupleResponse {
  id: string;
  inviteCode: string;
  partnerName: string;
  userId: string;
  startDate: string;
  currentStreak: number;
  longestStreak: number;
  isDemoMode: boolean;
}

export interface SessionResponse {
  userId: string;
  token: string;
  user: UserProfile;
  couple?: CoupleResponse;
}

export interface JoinCoupleRequest {
  /**
     * @minLength 6
     * @maxLength 6
     */
  inviteCode: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  userName: string;
}

export interface SaveEntryRequest {
  /** @maxLength 2000 */
  grateful: string;
  /** @maxLength 2000 */
  cute: string;
  /** @maxLength 2000 */
  grow: string;
  submitted: boolean;
  /** @maxLength 40 */
  reaction?: string;
}

export interface SharedEntry {
  date: string;
  grateful: string;
  cute: string;
  grow: string;
  submitted: boolean;
  partnerGrateful: string;
  partnerCute: string;
  partnerGrow: string;
  partnerSubmitted: boolean;
  partnerReaction?: string;
  myReaction?: string;
  revealed: boolean;
}