# lib/api-zod/src/generated/types/coupleResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CoupleResponse {
  id: string;
  inviteCode: string;
  partnerName: string;
  userId: string;
  startDate: string;
  currentStreak: number;
  longestStreak: number;
  isDemoMode: boolean;
}