# lib/api-zod/src/generated/types/createCoupleRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCoupleRequest {
  /**
     * @minLength 1
     * @maxLength 80
     */
  userName: string;
}