# lib/api-zod/src/generated/types/createSessionRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSessionRequest {
  /**
     * @minLength 8
     * @maxLength 40
     */
  accountKey: string;
  /**
     * @minLength 6
     * @maxLength 64
     */
  recoveryPin: string;
  /** @maxLength 80 */
  userName?: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  birthday?: string;
  /** @maxLength 40 */
  pronouns?: string;
}