# lib/api-zod/src/generated/types/saveEntryRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SaveEntryRequest {
  /** @maxLength 2000 */
  grateful: string;
  /** @maxLength 2000 */
  cute: string;
  /** @maxLength 2000 */
  grow: string;
  submitted: boolean;
  /** @maxLength 40 */
  reaction?: string;
}