# lib/api-zod/src/generated/types/sessionResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoupleResponse } from './coupleResponse';
import type { UserProfile } from './userProfile';

export interface SessionResponse {
  userId: string;
  token: string;
  user: UserProfile;
  couple?: CoupleResponse;
}