# lib/api-zod/src/generated/types/sharedEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SharedEntry {
  date: string;
  grateful: string;
  cute: string;
  grow: string;
  submitted: boolean;
  partnerGrateful: string;
  partnerCute: string;
  partnerGrow: string;
  partnerSubmitted: boolean;
  partnerReaction?: string;
  myReaction?: string;
  revealed: boolean;
}